R25-g: 63/63 全量 PASS 首次完全收敛 + 039 真根因=移植单位错误族(元素差>>1 折半)
- 活体收口证伪 R25-e script+0xd 假说(展开期实测 +0xd=0; 三张表枚举无 X)
- 真根因: dc (longlong)a-(longlong)b>>1 = 字节差>>1=元素数; msvc 误写 (a-b)>>1 = 折半
  1) core_execline.c:831(dc:78350) %X% 未命中段拷贝折半
  2) core_b2d.c ParseCommandBlock 四处(dc:22565/22626/22653/22675):
     TEAM 段 'CALC X=2*3' 拷成 'CALC ' → CALC 不写 X → FIND miss → exit 2
- 修复后 039 PASS(exit=0=golden); 全量 63/63 零回归(md5=fa90cb9c)
- 新立工单: >>1 单位错误族全库审计(语料盲区需逐点核 dc)
- 收敛曲线: 9(R19)→41(R24b)→53(R24e)→54→59→62→63/63(R25g)

+++ src/lang/core_execline.c
@@ -828,7 +828,10 @@ int64_t PECMD_ExpandEnvVars(void *script, WCHAR *line, WCHAR **out, int mode, ui
                 }
                 local_230 = (int64_t)(int)(char)xflag;
                 local_268 = '\0';
-                iVar10 = (int64_t)(p13 - p) >> 1;
+                /* dc:78350 (longlong)pWVar13-(longlong)pWVar8>>1 = 字节差>>1 ≡ 元素数;
+                 * 旧移植 (p13-p)>>1 把元素差再折半 → %X% 未命中段只拷 '%'(应拷 '%X'),
+                 * 039 活体定案: TEAM 行展开 '%X%'→'%', FIND $%=6 判假 exit=2 */
+                iVar10 = (int64_t)((intptr_t)p13 - (intptr_t)p) >> 1;
                 if (*p13 == L'%') {
                     nameStart = p11;
                     if (neg || (*p11 != L'&' && (xflag == 0 || *p11 == L'^'))) {
